+++ packages/dapp/.eslintrc.js
@@ -4,7 +4,7 @@ module.exports = {
     es2021: true,
     jest: true,
   },
-  extends: ['../../.eslintrc.js'],
+  extends: ['../../.eslintrc.js', 'plugin:@next/next/recommended'],
   rules: {
     'no-use-before-define': 'off',
     'import/no-default-export': 'off',