+++ packages/dapp/pages/invoice/[chainId]/[invoiceId]/instant.tsx
@@ -19,7 +19,7 @@ import {
 } from '@smartinvoicexyz/ui';
 import _ from 'lodash';
 import { useParams } from 'next/navigation';
-import { Address, formatUnits, Hex, hexToNumber, isAddress } from 'viem';
+import { Address, formatUnits, Hex, isAddress } from 'viem';
 import { useAccount, useChainId } from 'wagmi';
 
 import { useOverlay } from '../../../../contexts/OverlayContext';
@@ -29,14 +29,16 @@ function ViewInstantInvoice() {
     chainId: Hex;
     invoiceId: Address;
   }>();
-  const invoiceChainId = hexChainId && hexToNumber(hexChainId);
+  const invoiceChainId = hexChainId
+    ? parseInt(String(hexChainId), 16)
+    : undefined;
   const { modals, setModals } = useOverlay();
   const chainId = useChainId();
   const { address } = useAccount();
 
   const { invoiceDetails, isLoading } = useInvoiceDetails({
     address: invoiceId,
-    chainId: invoiceChainId,
+    chainId,
   });
 
   const {
@@ -79,8 +81,11 @@ function ViewInstantInvoice() {
 
   if (!invoiceDetails || isLoading) {
     return (
-      <Container overlay>
+      <Container overlay gap={10}>
         <Loader size="80" />
+        If the invoice does not load,
+        <br />
+        please refresh the browser.
       </Container>
     );
   }
@@ -209,10 +214,10 @@ function ViewInstantInvoice() {
             >
               <Text>Deposited</Text>
 
-              <Text>{`(${formatUnits(
+              <Text>{`${formatUnits(
                 amountFulfilled || BigInt(0),
                 tokenBalance?.decimals || 18,
-              )} ${tokenBalance?.symbol})`}</Text>
+              )} ${tokenBalance?.symbol}`}</Text>
             </Flex>
 
             <Divider