+++ packages/graphql/src/client.ts
@@ -2,18 +2,18 @@ import {
   ApolloClient,
   InMemoryCache,
   NormalizedCacheObject,
+  QueryOptions,
 } from '@apollo/client';
 import { SUPPORTED_NETWORKS } from '@smartinvoicexyz/constants';
 import { getGraphUrl } from '@smartinvoicexyz/shared';
 
+import { scalars } from './scalars';
 import {
   decodeScalarsInResponse,
   GenericOperation,
   GraphQLTypes,
   InputType,
   OperationOptions,
-  ScalarDefinition,
-  ThunderGraphQLOptions,
   ValueTypes,
   VType,
 } from './zeus';
@@ -33,44 +33,39 @@ const clients = SUPPORTED_NETWORKS.reduce(
   {} as Record<number, ApolloClient<NormalizedCacheObject>>,
 );
 
+type Scalars = typeof scalars;
+
 export const fetchTypedQuery =
   <
     O extends keyof typeof Ops,
-    SCLR extends ScalarDefinition,
     R extends keyof ValueTypes = GenericOperation<O>,
   >(
     chainId: number,
-    operation: O,
-    options?: ThunderGraphQLOptions<SCLR>,
   ) =>
   async <Z extends ValueTypes[R]>(
     o: Z & {
       [P in keyof Z]: P extends keyof ValueTypes[R] ? Z[P] : never;
     },
-    ops?: OperationOptions & { variables?: Record<string, unknown> },
-  ): Promise<InputType<GraphQLTypes[R], Z, SCLR>> => {
-    const gql = typedGql<O, SCLR, R>(operation, options)(o, ops);
-
-    if (operation === 'subscription') {
-      throw new Error('Subscriptions are not supported');
-    }
+    ops?: OperationOptions & { variables?: Record<string, unknown> } & Omit<
+        QueryOptions,
+        'query' | 'variables'
+      >,
+  ): Promise<InputType<GraphQLTypes[R], Z, Scalars>> => {
+    const gql = typedGql<O, Scalars, R>('query' as O, { scalars })(o, ops);
 
     const { data, error } = await clients[chainId].query({
       query: gql,
-      variables: ops?.variables,
+      ...ops,
     });
 
     if (error) throw error;
 
-    if (options?.scalars) {
-      return decodeScalarsInResponse({
-        response: data,
-        initialOp: operation,
-        initialZeusQuery: o as VType,
-        returns: ReturnTypes,
-        scalars: options.scalars,
-        ops: Ops,
-      });
-    }
-    return data as InputType<GraphQLTypes[R], Z, SCLR>;
+    return decodeScalarsInResponse({
+      response: data,
+      initialOp: 'query' as O,
+      initialZeusQuery: o as VType,
+      returns: ReturnTypes,
+      scalars,
+      ops: Ops,
+    });
   };