+++ packages/graphql/src/index.ts
@@ -2,4 +2,5 @@ export * from './client';
 export * from './invoice';
 export * from './invoices';
 export * from './scalars';
+export * from './sync';
 export * from './zeus';