+++ packages/hooks/src/useDeposit.ts
@@ -1,5 +1,5 @@
 import { IERC20_ABI, PAYMENT_TYPES, TOASTS } from '@smartinvoicexyz/constants';
-import { InvoiceDetails } from '@smartinvoicexyz/graphql';
+import { InvoiceDetails, waitForSubgraphSync } from '@smartinvoicexyz/graphql';
 import { UseToastReturn } from '@smartinvoicexyz/types';
 import { errorToastHandler } from '@smartinvoicexyz/utils';
 import { SimulateContractErrorType, WriteContractErrorType } from '@wagmi/core';
@@ -13,9 +13,6 @@ import {
   useSimulateContract,
   useWriteContract,
 } from 'wagmi';
-
-import { usePollSubgraph } from './usePollSubgraph';
-import { useTokenBalance } from './useTokenMetadata';
 
 export const useDeposit = ({
   invoice,
@@ -41,25 +38,9 @@ export const useDeposit = ({
 } => {
   const chainId = useChainId();
 
-  const { token, tokenBalance } = _.pick(invoice, ['token', 'tokenBalance']);
+  const { token } = _.pick(invoice, ['token', 'tokenBalance']);
 
   const publicClient = usePublicClient();
-
-  const { data: networkTokenBalance } = useTokenBalance({
-    address: invoice?.address as Hex,
-    tokenAddress: token as Hex,
-    chainId,
-  });
-
-  const waitForIndex = usePollSubgraph({
-    label: 'useDeposit',
-    fetchHelper: async () => {},
-    checkResult: () =>
-      tokenBalance?.value && networkTokenBalance
-        ? networkTokenBalance === amount + tokenBalance.value
-        : false,
-    interval: 2000, // 2 seconds, averaging about 20 seconds for index by subgraph
-  });
 
   const {
     data,
@@ -84,10 +65,12 @@ export const useDeposit = ({
     mutation: {
       onSuccess: async hash => {
         toast.info(TOASTS.useDeposit.waitingForTx);
-        await publicClient?.waitForTransactionReceipt({ hash });
+        const receipt = await publicClient?.waitForTransactionReceipt({ hash });
 
         toast.info(TOASTS.useDeposit.waitingForIndex);
-        await waitForIndex();
+        if (receipt && publicClient) {
+          await waitForSubgraphSync(publicClient.chain.id, receipt.blockNumber);
+        }
 
         onTxSuccess?.();
       },
@@ -98,10 +81,12 @@ export const useDeposit = ({
     mutation: {
       onSuccess: async hash => {
         toast.info(TOASTS.useDeposit.waitingForTx);
-        await publicClient?.waitForTransactionReceipt({ hash });
+        const receipt = await publicClient?.waitForTransactionReceipt({ hash });
 
         toast.info(TOASTS.useDeposit.waitingForIndex);
-        await waitForIndex();
+        if (receipt && publicClient) {
+          await waitForSubgraphSync(publicClient.chain.id, receipt.blockNumber);
+        }
 
         onTxSuccess?.();
       },