+++ packages/hooks/src/useInstantCreate.ts
@@ -4,7 +4,7 @@ import {
   TOASTS,
   wrappedNativeToken,
 } from '@smartinvoicexyz/constants';
-import { fetchInvoice, Invoice } from '@smartinvoicexyz/graphql';
+import { waitForSubgraphSync } from '@smartinvoicexyz/graphql';
 import { UseToastReturn } from '@smartinvoicexyz/types';
 import {
   errorToastHandler,
@@ -18,7 +18,7 @@ import { Address, encodeAbiParameters, Hex, parseUnits, toHex } from 'viem';
 import { usePublicClient, useSimulateContract, useWriteContract } from 'wagmi';
 
 import { useDetailsPin } from './useDetailsPin';
-import { usePollSubgraph } from './usePollSubgraph';
+import { useFetchTokens } from './useFetchTokens';
 
 export const useInstantCreate = ({
   invoiceForm,
@@ -39,14 +39,12 @@ export const useInstantCreate = ({
 } => {
   const invoiceFactory = getInvoiceFactoryAddress(chainId);
   const [waitingForTx, setWaitingForTx] = useState(false);
-  const [newInvoiceId, setNewInvoiceId] = useState<Hex | undefined>();
   const { getValues } = invoiceForm;
   const invoiceValues = getValues();
   const {
     client,
     provider,
     token,
-    tokenMetadata,
     projectName,
     projectDescription,
     projectAgreement,
@@ -60,7 +58,6 @@ export const useInstantCreate = ({
     'client',
     'provider',
     'token',
-    'tokenMetadata',
     'projectName',
     'projectDescription',
     'projectAgreement',
@@ -72,23 +69,21 @@ export const useInstantCreate = ({
     'lateFeeTimeInterval',
   ]);
 
-  const detailsData = {
-    projectName,
-    projectDescription,
-    projectAgreement: _.get(_.first(projectAgreement), 'src', ''),
-    startDate,
-    endDate,
-  };
+  const detailsData = useMemo(
+    () => ({
+      projectName,
+      projectDescription,
+      projectAgreement: _.get(_.first(projectAgreement), 'src', ''),
+      startDate,
+      endDate,
+    }),
+    [projectName, projectDescription, projectAgreement, startDate, endDate],
+  );
 
-  const { data: details } = useDetailsPin({ ...detailsData });
+  const { data: tokens } = useFetchTokens();
+  const tokenMetadata = _.filter(tokens, { address: token, chainId })[0];
 
-  const waitForResult = usePollSubgraph({
-    label: 'Creating escrow invoice',
-    fetchHelper: () =>
-      newInvoiceId ? fetchInvoice(chainId, newInvoiceId) : undefined,
-    checkResult: (v: Partial<Invoice>) => !_.isUndefined(v),
-    interval: 2000, // 2 seconds (averaging ~20 seconds for the subgraph to index)
-  });
+  const { data: details } = useDetailsPin(detailsData);
 
   const paymentAmount = useMemo(() => {
     if (!tokenMetadata || !paymentDue) {
@@ -186,10 +181,11 @@ export const useInstantCreate = ({
           'invoice',
         );
         if (!localInvoiceId) return;
-        setNewInvoiceId(localInvoiceId);
         toast.info(TOASTS.useInvoiceCreate.waitingForIndex);
 
-        await waitForResult();
+        if (txData && publicClient) {
+          await waitForSubgraphSync(publicClient.chain.id, txData.blockNumber);
+        }
         setWaitingForTx(false);
 
         // pass back to component for further processing