+++ packages/ui/src/forms/DatePicker.tsx
@@ -13,9 +13,7 @@ import {
 } from '@chakra-ui/react';
 import { getDateString } from '@smartinvoicexyz/utils';
 import _ from 'lodash';
-import ReactDatePicker, {
-  DatePickerProps as ReactDatePickerProps,
-} from 'react-datepicker';
+import ReactDatePicker from 'react-datepicker';
 import { Controller, RegisterOptions, UseFormReturn } from 'react-hook-form';
 
 // TODO handle separate controlled component
@@ -32,7 +30,7 @@ export type DatePickerProps = {
   placeholder?: string;
   variant?: string;
   spacing?: number | string;
-} & Omit<ReactDatePickerProps, 'onChange'>;
+} & Omit<React.ComponentProps<typeof ReactDatePicker>, 'onChange'>;
 
 export function DatePicker({
   label,