+++ packages/ui/src/organisms/VerifyInvoice.tsx
@@ -35,7 +35,6 @@ export function VerifyInvoice({
   };
 
   const { writeAsync, isLoading } = useVerify({
-    invoice,
     address: validAddress,
     chainId,
     toast,