+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @smartinvoicexyz/graphql
 
+## 0.1.6
+
+### Patch Changes
+
+- bug fixes
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.4
+  - @smartinvoicexyz/shared@0.1.4
+
+## 0.1.5
+
+### Patch Changes
+
+- fixed types
+
 ## 0.1.4
 
 ### Patch Changes