+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @smartinvoicexyz/hooks
 
+## 0.1.9
+
+### Patch Changes
+
+- bug fixes
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.4
+  - @smartinvoicexyz/graphql@0.1.6
+  - @smartinvoicexyz/types@0.1.6
+  - @smartinvoicexyz/utils@0.1.7
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @smartinvoicexyz/graphql@0.1.5
+  - @smartinvoicexyz/types@0.1.5
+  - @smartinvoicexyz/utils@0.1.6
+
 ## 0.1.7
 
 ### Patch Changes