+++ packages/types/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @smartinvoicexyz/types
 
+## 0.1.6
+
+### Patch Changes
+
+- bug fixes
+- Updated dependencies
+  - @smartinvoicexyz/graphql@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @smartinvoicexyz/graphql@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes