+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @smartinvoicexyz/utils
 
+## 0.1.7
+
+### Patch Changes
+
+- bug fixes
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.4
+  - @smartinvoicexyz/graphql@0.1.6
+  - @smartinvoicexyz/types@0.1.6
+  - @smartinvoicexyz/shared@0.1.4
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @smartinvoicexyz/graphql@0.1.5
+  - @smartinvoicexyz/types@0.1.5
+
 ## 0.1.5
 
 ### Patch Changes